# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  polling-system_auth_models.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      firebase_uid:
        type: string
      id:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
...
  polling-system_polling_models.Answer:
    properties:
      answer:
        type: string
      candidate:
        $ref: '#/definitions/polling-system_polling_models.Candidate'
      candidate_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      question:
        $ref: '#/definitions/polling-system_polling_models.PollQuestion'
      question_id:
        type: string
      source:
        type: string
    type: object
...
  polling-system_polling_models.BulkCandidateCreate:
    properties:
      phone_numbers:
        items:
          type: string
        minItems: 1
        type: array
    required:
    - phone_numbers
    type: object
  polling-system_polling_models.BulkCandidateResult:
    properties:
      created:
        type: integer
      duplicates:
        items:
          type: string
        type: array
      invalid:
        items:
          type: string
        type: array
    type: object
  polling-system_polling_models.Candidate:
    properties:
      answers:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      created_at:
        type: string
      id:
        type: string
      phone:
        type: string
      poll:
        $ref: '#/definitions/polling-system_polling_models.Poll'
      poll_id:
        type: string
      updated_at:
        type: string
      voice_poll_sentiment:
        type: string
      voice_poll_started_at:
        type: string
      voicemail_detected:
        type: boolean
    type: object
  polling-system_polling_models.CandidateCreate:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  polling-system_polling_models.CandidateUpdate:
    properties:
      phone:
        type: string
      voice_poll_sentiment:
        type: string
      voice_poll_started_at:
        type: string
      voicemail_detected:
        type: boolean
    type: object
  polling-system_polling_models.ErrorResponse:
    properties:
      error:
        example: something went wrong
        type: string
    type: object
...
  polling-system_polling_models.PaginatedCandidates:
    properties:
      items:
        items:
          $ref: '#/definitions/polling-system_polling_models.Candidate'
        type: array
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
...
  polling-system_polling_models.Poll:
    properties:
      bland_pathway_ids:
        type: object
      candidates:
        items:
          $ref: '#/definitions/polling-system_polling_models.Candidate'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/polling-system_auth_models.User'
      custom_voice_poll_instructions:
        type: string
      dont_send_sms:
        type: boolean
      first_typeform_sms_text:
        type: string
      id:
        type: string
      only_day_hours:
        type: boolean
      poll_name:
        type: string
      poller_name:
        type: string
      questions:
        items:
          $ref: '#/definitions/polling-system_polling_models.PollQuestion'
        type: array
      scheduled_to_start_at:
        type: string
      status:
        type: string
      time_zone:
        type: string
      typeform_forms:
        type: object
      updated_at:
        type: string
      voice_poll_delay_minutes:
        type: integer
      voice_poll_method:
        type: string
    type: object
...
  polling-system_polling_models.PollQuestion:
    properties:
      answer_options:
        type: object
      answer_type:
        type: string
      answers:
        items:
          $ref: '#/definitions/polling-system_polling_models.Answer'
        type: array
      created_at:
        type: string
      default_next_question:
        $ref: '#/definitions/polling-system_polling_models.PollQuestion'
      default_next_question_id:
        type: string
      flow_rules:
        type: object
      id:
        type: string
      poll:
        $ref: '#/definitions/polling-system_polling_models.Poll'
      poll_id:
        type: string
      question:
        type: string
      randomize_answers:
        type: boolean
      typeform_field_id:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Unified API for authentication, polls, contacts, and file management.
  title: Polling System API
  version: "1.0"
paths:
...
  /candidates:
    get:
      description: Returns a paginated list of candidates, filtered by poll_id.
      parameters:
      - description: Poll UUID to filter by
        in: query
        name: poll_id
        required: true
        type: string
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PaginatedCandidates'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Query candidates
      tags:
      - Candidates
...
  /candidates/{id}:
    delete:
      description: Deletes a candidate and its answers. Requires ownership of the
        parent poll for pollers.
      parameters:
      - description: Candidate UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a candidate
      tags:
      - Candidates
    get:
      description: Returns a single candidate.
      parameters:
      - description: Candidate UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.Candidate'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get candidate by ID
      tags:
      - Candidates
    patch:
      consumes:
      - application/json
      description: Partially updates a candidate. Requires ownership of the parent
        poll for pollers.
      parameters:
      - description: Candidate UUID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.CandidateUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.Candidate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a candidate
      tags:
      - Candidates
...
  /polls/{pollId}/candidates:
    get:
      description: Returns a paginated list of candidates for the given poll.
      parameters:
      - description: Poll UUID
        in: path
        name: pollId
        required: true
        type: string
      - default: 1
        description: 'Page number (min: 1)'
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (1-100)
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling-system_polling_models.PaginatedCandidates'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List candidates for a poll
      tags:
      - Candidates
    post:
      consumes:
      - application/json
      description: Adds a candidate to a poll. Requires poll ownership for pollers.
      parameters:
      - description: Poll UUID
        in: path
        name: pollId
        required: true
        type: string
      - description: Candidate to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.CandidateCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_polling_models.Candidate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add candidate to poll
      tags:
      - Candidates
  /polls/{pollId}/candidates/bulk:
    post:
      consumes:
      - application/json
      description: Creates candidates from a list of phone numbers. Deduplicates and
        normalizes phones.
      parameters:
      - description: Poll UUID
        in: path
        name: pollId
        required: true
        type: string
      - description: Phone numbers to add
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/polling-system_polling_models.BulkCandidateCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_polling_models.BulkCandidateResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk add candidates to a poll
      tags:
      - Candidates
  /polls/{pollId}/candidates/upload:
    post:
      consumes:
      - multipart/form-data
      description: Parses a CSV file with a 'phone' column and creates candidates
        for the poll. Deduplicates and normalizes phones.
      parameters:
      - description: Poll UUID
        in: path
        name: pollId
        required: true
        type: string
      - description: CSV file with phone column (.csv)
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/polling-system_polling_models.BulkCandidateResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/polling-system_polling_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Upload CSV to bulk add candidates
      tags:
      - Candidates
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"